Implement install precheck
