+++ main.go
@@ -4,11 +4,11 @@ import (
 	"fmt"
 	"os"
 
-	"github.com/everettraven/packageless/utils"
+	"github.com/everettraven/packageless/subcommands"
 )
 
 func main() {
-	if err := utils.SubCommand(os.Args[1:]); err != nil {
+	if err := subcommands.SubCommand(os.Args[1:]); err != nil {
 		fmt.Println(err)
 		os.Exit(1)
 	}
